requirements/base.txt
analytics-python==1.1.0
django==1.10.7
django-compressor==2.1.1
django-crispy-forms==1.6.1
django_extensions==1.5.5
django-filter==1.0.1
django-libsass==0.5
django-oscar==1.4
django-rest-swagger[reST]==0.3.10
django-simple-history==1.8.2
django-solo==1.1.2
django-threadlocals==0.8
django-waffle==0.12.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
drf-extensions==0.3.1
edx-auth-backends==1.1.2
edx-django-release-util==0.3.0
edx-django-sites-extensions==1.0.0
edx-drf-extensions==1.2.2
edx-ecommerce-worker==0.6.0
edx-opaque-keys==0.3.1
edx-rest-api-client==1.7.1
jsonfield==1.0.3
libsass==0.9.2
ndg-httpsclient==0.4.2
path.py==7.2
paypalrestsdk==1.11.5
premailer==2.9.2
pycountry==17.1.8
python-dateutil==2.4.2
pytz==2016.10
requests==2.17.3
sailthru-client==2.2.3
six==1.10.0
zeep==2.1.1
spgateway-core==0.1.1
